+++ CHANGES.md
@@ -22,13 +22,16 @@
 ## 2012-Nov-?? / 0.2.0
 
 * Project config support (`:clr` key in `project.clj`)
-  * Support for `:assembly-paths` (transparently calls `assembly-load-from`)
+  * Support for `:assembly-paths`
+    * transparently calls `assembly-load-from` for matching assemblies
   * Command resolution
     * Support searching within env-var value, e.g. `[*PATH "foo.exe"]`
     * Command templates `:cmd-templates`
 * Dependency support
   * [TODO] Maven dependencies on CLOJURE_LOAD_PATH
-  * Assembly dependencies (via NuGet etc.) -- `:deps-cmds` `:deps-regex`
+  * Command-based dependencies (via NuGet etc.) -- `:deps-cmds`
+  * Assembly deps regex `:assembly-deps-regex` to match versions/types
+    * transparently calls `assembly-load-from` for matching assemblies
 * Fixes for `compile` task
 * [TODO] sample.project.clj with config options